requirements.txt
# Research: Bidirectional Validation of Molecular Reasoning Structures
# Install: pip install -r requirements.txt

# Core API
anthropic>=0.40.0

# Sentence embeddings for bond classification
sentence-transformers>=2.2.0

# TDC benchmark data
PyTDC>=1.0

# Statistical analysis
scipy>=1.12.0
scikit-learn>=1.4.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Data
pandas>=2.1.0
numpy>=1.26.0

# Molecule handling (rdkit already installed via chemprop)
# rdkit>=2023.9.0

# HTTP client (for tool execution against quanta-mcp)
httpx>=0.27.0

# HuggingFace datasets (for OpenThoughts validation)
datasets>=2.18.0

# Utilities
tqdm>=4.66.0
